requirements.txt
requests
apscheduler 
python-dotenv
supabase
google-genai
# apscheduler